+++ packages/at_onboarding_cli/example/apkam_examples/enroll_app_listen.dart
@@ -56,10 +56,16 @@ Future<void> _notificationCallback(AtNotification notification,
   var enrollParamsJson = {};
   enrollParamsJson['enrollmentId'] = enrollmentId;
   if (approveResponse == 'yes') {
-    final encryptedApkamSymmetricKey =
-        jsonDecode(notification.value!)['encryptedApkamSymmetricKey'];
+    var encryptedAPKAMSymmetricKey =
+        jsonDecode(notification.value!)['encryptedAPKAMSymmetricKey'];
+    // workaround for a server issue where it may send encryptedAPKAMSymmetricKey or encryptedApkamSymmetricKey
+    if (encryptedAPKAMSymmetricKey == null ||
+        encryptedAPKAMSymmetricKey.isEmpty) {
+      encryptedAPKAMSymmetricKey =
+          jsonDecode(notification.value!)['encryptedApkamSymmetricKey'];
+    }
     final apkamSymmetricKey = EncryptionUtil.decryptKey(
-        encryptedApkamSymmetricKey, atAuthKeys.defaultEncryptionPrivateKey!);
+        encryptedAPKAMSymmetricKey, atAuthKeys.defaultEncryptionPrivateKey!);
     print('decrypted apkam symmetric key: $apkamSymmetricKey');
     var encryptedDefaultPrivateEncKey = EncryptionUtil.encryptValue(
         atAuthKeys.defaultEncryptionPrivateKey!, apkamSymmetricKey);