+++ packages/at_onboarding_cli/lib/src/util/at_onboarding_preference.dart
@@ -1,6 +1,5 @@
 import 'dart:core';
 
-import 'package:at_chops/at_chops.dart';
 import 'package:at_client/at_client.dart';
 import 'package:at_onboarding_cli/src/util/registrar_api_constants.dart';
 
@@ -11,14 +10,6 @@ class AtOnboardingPreference extends AtClientPreference {
   /// specify path of qr code containing cram secret
   @Deprecated('qr_code based cram authentication not supported anymore')
   String? qrCodePath;
-
-  /// signing algorithm to use for pkam authentication
-  @override
-  SigningAlgoType signingAlgoType = SigningAlgoType.rsa2048;
-
-  /// hashing algorithm to use for pkam authentication
-  @override
-  HashingAlgoType hashingAlgoType = HashingAlgoType.sha256;
 
   PkamAuthMode authMode = PkamAuthMode.keysFile;
 
@@ -34,5 +25,6 @@ class AtOnboardingPreference extends AtClientPreference {
 
   String? deviceName;
 
+  @Deprecated("No longer used")
   int apkamAuthRetryDurationMins = 30;
 }