+++ packages/at_commons/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 4.0.10
+- fix: Deprecate apkam in PkamAuthMode enum
 - chore: deprecate MessageTypeEnum.text
 ## 4.0.9
 - feat: enroll verb syntax change for enroll:revoke:force and added new exception AtEnrollmentRevokeException